+++ playwright/playwright-utils.ts
@@ -1,14 +1,7 @@
-import { Locator, Page } from '@playwright/test'
-
-export const kirjauduVirkailijanNakymaan = async (
-  page: Page,
-  ticket?: string
-): Promise<void> => {
-  await page.goto(`/lomake-editori/auth/cas?ticket=${ticket || 'DEVELOPER'}`)
-}
-
-export const getSensitiveAnswer = (page: Page): Locator =>
-  page.getByTestId('checkbox-sensitive-answer')
+import { Page, Response } from '@playwright/test'
+import * as Record from 'fp-ts/lib/Record'
+import * as Option from 'fp-ts/lib/Option'
+import { AssertionError } from 'assert'
 
 type HttpMethod = 'PUT' | 'POST' | 'GET' | 'DELETE'
 
@@ -20,3 +13,23 @@ export const waitForResponse = (
   page.waitForResponse((response) => {
     return response.request().method() === method && urlMatcher(response.url())
   })
+
+export const getJsonResponseKey = async <T>(res: Response, key: string) => {
+  try {
+    const body = await res.json()
+    return Record.hasOwnProperty(key, body)
+      ? Option.some(body[key] as T)
+      : Option.none
+  } catch (e) {
+    return Option.none
+  }
+}
+
+export const unsafeFoldOption = <T>(o: Option.Option<T>): T => {
+  return Option.fold<T, T>(
+    () => {
+      throw new AssertionError({ message: 'Option was None' })
+    },
+    (val) => val
+  )(o)
+}
